created notifications provider fixed list of notifications fixed icons size
